new testhelpers module to share test extensions
De-duplicated ContainedRobolectricRunner.kt and RobolectricExtension.kt
by moving them into their own module.

Before landing on this implementation I attempted to use the gradle
feature testFixtures, however this doesn't work with the android
gradle plugin.
