+++ OneSignalSDK/onesignal/core/src/test/java/com/onesignal/session/internal/outcomes/OutcomeEventsRepositoryTests.kt
@@ -2,7 +2,6 @@ package com.onesignal.session.internal.outcomes
 
 import com.onesignal.debug.LogLevel
 import com.onesignal.debug.internal.logging.Logging
-import com.onesignal.extensions.RobolectricTest
 import com.onesignal.mocks.DatabaseMockHelper
 import com.onesignal.session.internal.influence.Influence
 import com.onesignal.session.internal.influence.InfluenceChannel
@@ -13,6 +12,7 @@ import com.onesignal.session.internal.outcomes.impl.OutcomeEventsRepository
 import com.onesignal.session.internal.outcomes.impl.OutcomeEventsTable
 import com.onesignal.session.internal.outcomes.impl.OutcomeSource
 import com.onesignal.session.internal.outcomes.impl.OutcomeSourceBody
+import com.onesignal.testhelpers.extensions.RobolectricTest
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
