+++ OneSignalSDK/settings.gradle
@@ -25,3 +25,4 @@ include ':OneSignal:core'
 include ':OneSignal:in-app-messages'
 include ':OneSignal:location'
 include ':OneSignal:notifications'
+include ':OneSignal:testhelpers'
